style(ppt): clear the clippy backlog in legacy powerpoint code
- split same-line if chains in the animation editor
- drop explicit into_iter() on an array argument
- enumerate() instead of a manual master-ordinal counter
- remove no-op shift/cast; to_bytes takes self by value (Copy type)
- targeted allows: large_enum_variant on the public envelope payload
  enum, type_complexity on the OLE-children parser

ole/ppt now reports zero clippy warnings; litchi-ole suite passes

+++ crates/litchi-ole/src/ppt/text_si_exception.rs
@@ -154,7 +154,7 @@ impl PowerPointTextSpecialInfoDefaults {
         data.extend_from_slice(&0u16.to_le_bytes());
         data.extend_from_slice(&TEXT_SPECIAL_INFO_DEFAULT_TYPE.to_le_bytes());
         data.extend_from_slice(
-            &((4 + 2 * mask.count_ones()) as u32).to_le_bytes(),
+            &(4 + 2 * mask.count_ones()).to_le_bytes(),
         );
         data.extend_from_slice(&mask.to_le_bytes());
         if let Some(spelling) = self.spelling {
